Add files via upload
This program attempts to solve the trolley problem were given number of input for passengers and pedestrians and given conditions whether the traffic signal is green or red, it gives user choice to save either the passengers or the pedestrians when a car goes out of control also given the parameters such as the age of the passengers or pedestrians, a number of passengers or pedestrians etc. makes an automated decision to save either passengers or pedestrians.